Text-Scroll examples for Sharp 1403-series
This is a multi-type-scroll code example for Sharp 1403-series. Feel free to use this code. Scrolling line by line from left to right, from right to left and random.
I planned first to use this code example when the 1403H will arrive, but as it will take one or two weeks more, I published it using PocketEmul.
First it will create an array with the correct display adresses as they are not linear. Then it will display the text (24 chars) that should be slided. Then it reads the peek values from the display and finally it clears the screen and starts to write the stored values line by line depending on which below mentioned code you use. The process takes some time, so please be patient.

Why the non-linear to linear conversion:
Adress on 1403 are like these:
12288-12317  Display from 1-6
12333-12347  Display from 7-9
12318-12332  Display from 10-12
12382-12396  Display from 15-13
12397-12411  Display from 18-16
12352-12381  Display from 24-19

If you change this line in SCROLL1.BAS as showed below, scrolling will be from right to left:

370 FOR R = 120 TO 1 STEP -1

Instead, SCROLL2.BAS shows lines in random order.