PETOOLS System Extension
For the PC-1403H and the PC-1403.
PETOOLS 1.1 is a set of machine language routines, which extends the possibilities of your sharp in graphics, sound and other functions. It comes with a GPL and without warranty of any kind. With PETOOLS 1.1 it is possible, to write powerful basic programs easier. It is also easier, to port programs from the PC-1500 to use them with the PC-1403(H).
API documentation includes a description of every machine language routine and also instructions how to build PETOOLS 1.1